refactor: change file name main page to search page
